move version package to pkg/x/version
