+++ internal/server/get_version_info_v1alpha1.go
@@ -5,8 +5,8 @@ import (
 
 	"connectrpc.com/connect"
 
-	"github.com/akuity/kargo/internal/version"
 	svcv1alpha1 "github.com/akuity/kargo/pkg/api/service/v1alpha1"
+	"github.com/akuity/kargo/pkg/x/version"
 )
 
 func (s *server) GetVersionInfo(
